Add XCTAssertApodiniApplicationNotRunning
