+++ Sources/XCTApodini/XCTApodiniTest.swift
@@ -11,6 +11,7 @@ open class XCTApodiniTest: XCTestCase {
     // swiftlint:disable implicitly_unwrapped_optional
     open var app: Application!
     
+    
     override open func setUpWithError() throws {
         try super.setUpWithError()
         app = Application()
@@ -20,25 +21,7 @@ open class XCTApodiniTest: XCTestCase {
         try super.tearDownWithError()
         app.shutdown()
         
-        let processesAtPort8080 = runShellCommand(.getProcessesAtPort(8080))
-        if !processesAtPort8080.isEmpty {
-            #if !os(Linux)
-            XCTFail(
-                """
-                A web service is running at port 8080 after running the test case.
-                All processes at port 8080 must be shut down after running the test case.
-                """
-            )
-            #else
-            print(
-                """
-                A web service is running at port 8080 after running the test case:
-                    \(runShellCommand(.getProcessesAtPort(8080)))
-                """
-            )
-            #endif
-            runShellCommand(.killPort(8080))
-        }
+        XCTAssertApodiniApplicationNotRunning()
     }
     
     open func database() throws -> Database {
